Added Spider AI and Player Death
Added logic for the spider to chase you once you get close, and if it
catches you, you die.  On death there is a brief loss of control of your
character, then you respawn.  Spider's speed is based on proximity to
the player, the closer you get, the faster it chases you.